Stage 1: Conception, pre prototype phase
Wanted to be able to prop up any board & have space beneath. I realized the biggest things between me and any prototypes were all in the corners. "Tubing" in the walls? They need holes. The cylinders themselves? Use chopsticks. The board design? Cardboard and a pencil. Why would you start with trying to build the housing if you don't know what size board is reasonable, what properties it might have (such as needing space beneath it), or what the board should be made of? Let's start with any parts I won't be rapidly going through. One thing is certain, it has 4 corners, so let's pre fabricate that part and all else will easily follow. Most of the funny engineering either goes in the corners or on something that goes in the corners. The corners are designed to be tall enough to encompass anything that might fit underneath, wide enough to provide stability and give the cylinders some space, and short enough to fit the smallest size board I'll reasonably ever test, which I estimate to be about 3 square inches or 20 square cm. It has material where I'm not sure if it should be there because I can always chisel away what's in the way. 
"It's amazing what we could get done if we didn't care who got the credit" - Harry S. Truman. "You can't keep snakes in your backyard and expect them only to bite your neighbors" - Hillary Clinton (as US Secretary of State).
The world is a complicated place.